Change LSH implementation
Previously we were using Random Fourier Features embedding +
HyperplaneLSH. The combination was effective, but the complexity of
computing the hash values had a quadratic preprocessing term in the
length of the window, because of the embedding. Furthermore, the
subsequent dot products did not involve the original values of the time
series, making it difficult to use the Cyclical Convolution Theorem.

With this change we move to the (random projection LSH for p-stable
distributions)[http://theory.csail.mit.edu/~mirrokni/pstable.ps], which
opens the possibility of computing the dot products in the frequency
domain (still to be implemented).

While the new LSH implementation features one more parameter compared to
HyperplaneLSH (the width of the projection), the previous scheme needed
to set the scale factor of the embedding. Therefore we are trading one
parameter for another, which is not bad.