Save space by mapping LSH hash values to 32 bits.
Until commit `8739619558cd3bf34ef156f279f5766afff8171b` the hash values
were inserted in the hash matrix columns as-is, that is as collections
of 32 bytes. With this commit we are applying an additional level of
hashing (the standard one, not LSH) so to map long hash values to 32
bits integers. This way we save a factor 8 in memory, at the expense of
possibly more collisions, which can nonetheless be removed by means of
the deduplication machinery. Early experiments suggest that this might
also be faster.

Another advantage is that traversing the columns is much more
cache-friendly.

The drawback is that we have to sort the hashes every time we examine a
new level. But we do examine just two levels, so it might not be that
much of an issue.