use tensored failure probability
use tensored failure probability instead of the independent one,
as was done in the motifs VLDB paper. Alas, when looking for motiflets
we have to confirm distances that require an outrageous number of
independent repetitions to bring the tensored failure probability down enough.
In fact, it turns out that the failure probability assuming independence is
orders of magnitude smaller than the failure probability that takes into account
tensoring.

We will have to think about what to do in order to make the whole thing usable
for motiflets.