feat: simplify exploration of the index
This radically simplifies the navigation of the index during motiflet
discovery. Instead of trying to guess what would be the best prefix
to confirm th next distance to confirm (which might be a wildly off
guess of the actual extent of a motiflet), we now simply continue with
the next repetition on the current prefix, until there are no more
repetitions available and we have to consider shorter prefixes.

As an optimization, if we find out that using a shorter prefix
and _fewer_ repetitions than the ones computed so far would
lead to fewer collisions to confirm a candidate distance, then
we directly jump to said prefix.